fix(runtime): ExecClient keyed one-shot replay on ambiguous transport
Move guest-exec retry ownership into ExecClient::exec_command so legacy
CLI/SDK socket callers that mint request_id get in-call replay. Drop the
outer VmManager retry to avoid double attempts. Omit-path stays unkeyed.
